requirements.txt
streamlit==1.38.0
requests>=2.31.0
Pillow>=10.4.0
google-auth>=2.33.0
openai>=1.40.0
python-dotenv>=1.0.1
